[DEV-667] renamed ethereum blockchain directory
