+++ .env.example
@@ -35,6 +35,7 @@ UTXO_SPENDER_ADDRESS=
 ETH_WALLET_ADDRESS=
 ETH_WALLET_PRIVATE_KEY=
 ETH_GATEWAY_URL=
+ETH_API_KEY=
 ETH_EXPLORER_URL=
 LETTER_USER=
 LETTER_AUTH=
