+++ src/blockchain/shared/enums/blockchain.enum.ts
@@ -2,4 +2,5 @@ export enum Blockchain {
   DEFICHAIN = 'DeFiChain',
   BITCOIN = 'Bitcoin',
   ETHEREUM = 'Ethereum',
+  BINANCE_SMARTCHAIN = 'BinanceSmartchain',
 }
